demo/model_services/perceptron/requirements.txt
werkzeug
flask
tornado
requests
numpy
Pillow
opencv-python
